pyproject.toml
[project]
name = "mcp-prompt-server"
version = "0.1.0"
description = "MCP server that exposes markdown files as prompts"
readme = "README.md"
authors = [
    { name = "marc", email = "info@example.com" }
]
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.15.0",
    "pyyaml>=6.0",
]

[project.scripts]
mcp-prompt-server = "src.server:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
include = ["prompts", "resources"]

[tool.hatch.build.targets.wheel.force-include]
"prompts" = "prompts"
"resources" = "resources"
